Some Fix (Renaming, new "Rules" for export for "Eslint").
