+++ src/game/scripts/const/GameColor.ts
@@ -1,0 +1,3 @@
+export enum GameColor {
+  Example = '#ffffff',
+}
