--- src/game/scripts/const/project-color.ts
@@ -1,5 +1,0 @@
-enum ProjectColor {
-  ExampleColor = '#ffffff',
-}
-
-export default ProjectColor;
