+++ src/game/scripts/prefabs/CounterFPS.ts
