+++ src/game/scripts/prefabs/ExamplePrefab.ts
